+++ core_networking/src/main/java/com/example/core_networking/CurrencyResponse.kt
@@ -4,5 +4,5 @@ data class CurrencyResponse(
     val baseCurrency: String,
     val rates: Map<String, Double>
 ) {
-   val convertedRates: List<Rate> = rates.map { Rate(it.key, it.value) }
+   val convertedRates: List<Rate> = rates.map { Rate(it.key, it.value.toBigDecimal()) }
 }
