+++ core_networking/src/main/java/com/example/core_networking/Rate.kt
@@ -1,8 +1,10 @@
 package com.example.core_networking
 
+import java.math.BigDecimal
+
 data class Rate(
     val name : String,
-    val rate: Double,
+    val rate: BigDecimal,
     val isBaseCurrency: Boolean = false,
-    val convertedValue: Double? = null
+    val convertedValue: BigDecimal? = null
 )
